Feature/import excel

+++ SGULibraryManagement/BUS/DeviceBUS.cs
@@ -12,12 +12,21 @@ namespace SGULibraryManagement.BUS
     public class DeviceBUS
     {
         private readonly DeviceDAO DAO = new();
+        private readonly ReservationDAO reservationDAO = new();
+        private readonly BorrowDevicesDAO borrowDeviceDAO = new();
+
         private List<DeviceDTO> devices = [];
         public Dictionary<long, int> DeviceBorrowQuantity;
 
         public DeviceBUS()
         {
             DeviceBorrowQuantity = DAO.GetAllWithBorrowQuantity().ToDictionary(pair => pair.First.Id, pair => pair.Last);
+        }
+
+
+        public bool CreateListDevice(List<DeviceDTO> listDevice)
+        {
+            return DAO.CreateListDevice(listDevice);
         }
 
         public List<DeviceDTO> GetAll()
@@ -42,6 +51,14 @@ namespace SGULibraryManagement.BUS
 
         public bool Delete(DeviceDTO device)
         {
+            var borrows = borrowDeviceDAO.FindByDeviceId(device.Id);
+            var reservations = reservationDAO.FindByDeviceId(device.Id);
+
+            if (borrows.Count > 0 || reservations.Count > 0)
+            {
+                return false;
+            }
+
             return DAO.Delete(device.Id);
         }
